org.lockss.plugin.highwire.HighWireArchivalUnit *fixed int/long casting problem
org.lockss.daemon.Configuration
*changed getTimeIntervalParam(string, int) to (string, long)


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@831 4f837ed2-42f5-46e7-a7a5-fa17313484d4

+++ src/org/lockss/daemon/Configuration.java
@@ -1,5 +1,5 @@
 /*
- * $Id: Configuration.java,v 1.21 2003-03-05 21:07:14 tal Exp $
+ * $Id: Configuration.java,v 1.22 2003-03-08 01:44:32 troberts Exp $
  */
 
 /*
@@ -607,7 +607,7 @@ public abstract class Configuration {
   /** Static convenience method to get param from current configuration.
    * Don't accidentally use this on a <code>Configuration</code> instance.
    */
-  public static long getTimeIntervalParam(String key, int dfault) {
+  public static long getTimeIntervalParam(String key, long dfault) {
     return currentConfig.getTimeInterval(key, dfault);
   }
